pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testforge"
version = "0.2.0"
description = "CLI for generating tests, analyzing diffs, and validating regressions across branches."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "TestForge contributors" }]
keywords = ["cli", "testing", "git", "llm"]
dependencies = [
    "typer>=0.12.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
testforge = "testforge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["testforge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
